more refactoring, and documenting
