+++ src-tauri/Cargo.toml
@@ -1,7 +1,7 @@
 [package]
-name = "macos-task-manager"
+name = "neohtop"
 version = "1.1.0"
-description = "A Tauri App"
+description = "A cross-platform system monitor"
 authors = ["you"]
 edition = "2021"
