+++ src-tauri/src/ui/mod.rs
@@ -1,0 +1,7 @@
+//! User interface functionality
+//!
+//! This module handles UI-specific functionality, including window effects
+//! and platform-specific visual customizations.
+
+mod window;
+pub use window::setup_window_effects;
