--- src-tauri/src/window.rs
@@ -1,32 +1,0 @@
-#[cfg(target_os = "windows")]
-use window_vibrancy::apply_acrylic;
-#[cfg(target_os = "macos")]
-use window_vibrancy::{apply_vibrancy, NSVisualEffectMaterial, NSVisualEffectState};
-
-#[cfg(target_os = "windows")]
-pub fn setup_window_effects(
-    window: &tauri::WebviewWindow,
-) -> Result<(), Box<dyn std::error::Error>> {
-    apply_acrylic(window, Some((0, 0, 25, 125)))?;
-    Ok(())
-}
-
-#[cfg(target_os = "macos")]
-pub fn setup_window_effects(
-    window: &tauri::WebviewWindow,
-) -> Result<(), Box<dyn std::error::Error>> {
-    apply_vibrancy(
-        window,
-        NSVisualEffectMaterial::HudWindow,
-        Some(NSVisualEffectState::Active),
-        None,
-    )?;
-    Ok(())
-}
-
-#[cfg(not(any(target_os = "windows", target_os = "macos")))]
-pub fn setup_window_effects(
-    _window: &tauri::WebviewWindow,
-) -> Result<(), Box<dyn std::error::Error>> {
-    Ok(())
-}
